+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fixed issue with haptic feedback on comment actions
 - Fixed issue where reply button would cover actions and comments for short posts
 - Removed mark as read for mentions and replies that have already been read
+- Fixed issue where setting a default sort type would cause the app to infinitely load
 
 ## 0.2.1+8 - 2023-06-28
 ### Added